Login Form
I have created the login form with the input boxes in html for the username and password, also providing a login button for the user to click. I have also created a style.css file that holds all the decoration for the login form form only so far, more can be added to it for each individual page, as long as they have their own unique identifying names that will be used.