Edition of login form stylesheet
I have edited some of the detailing for the login form, making some small changes to the positioning of text whilst ensuring that it retains its professional look. I have also added in the required code comments for the stylesheet so we know where the code for the login form starts and ends.