Logging into System
I have now got the SQL behind the login page working. It now allows the user to log into the system. For now it redirects them to a index page that displays their name and allows them to log out, since this was to test it. I have implemented code that validates the users data before allowing them access, however, this is still underdevelopment. It has worked but I will still continue testing it to ensure it is consistent.