Login Attempt Limiter
I have implemented some code that starts a login attempt count and increases it after each failed attempt, upon the third attempt the user is then blocked for 30 seconds. Upon the first failed attempt alongside the original error message it also contains the number of attempts that the user has left. Once the timer has run through it should reset the counter, however this is still being tested and refined so isn't perfect.