Logo and label editing
I have added a class to the logo image on the login form so that it doesn't interfere with other images throughout the project. I have also edited the label placements of the login form to make sure that they are more central and have better spacing between the username inputs and the password input fields.