Registration Inputs
I have altered the styling of the inputs so all the placeholder labels are now central on the form. I have also removed and added the needed user inputs, such as their student number and a select dropdown box that allows them to pick the courses available. This dropdown menu will be edited to include the needed code so it shows the courses that are available in the database, not from a predetermined list.