Emails
I have implemented the code that will email the lecturer to approve of the accounts creation, it says that a message should appear on the lecturers dashboard according to the spec but since that page is not available an email will suffice for now. And I have added the additional functionality of should the user details be approved and added to the database, then they will get a confirmation email saying that they can access the test system.