Fixed password hashing`
The password hashing has now been fixed, when the data is inserted into the database it shows as the hashed version using Bcrypt. The problem was I needed to changed the order of the s's and i's in the prepared statements so they matched the order of the variables being used in the brackets next to it.