Fixing links to other pages
I have now altered my links throughout all my pages so that they coincide with the changes that were implemented with the routing. We have run into an issue where the includes connect file is not being seen by the code and this has been officially raised and will be addressed in due course.