Further landing edits
I have edited the routing for the landing page so that instead of having to put in the word landing, a slash is all that's needed since the user is going to the home page in a sense. Added in a margin between the two buttons so that it is more aesthetically pleasing.