Landing button styling
I have altered the colouring of the container of the landing page so it is conforming to the colours mentioned in the housestyle sheet. I have also changed the colours of the buttons so that when they hover they changed to the opposite colours, while staying with the coolours of the housestyle sheet.