OOP integration
I have altered the mysqli_real_escape_string so they reflect the required OOP format. I have also run a formatter on my files so that they are displayed in a clear and easy to understand layout. On top of that I have also adjusted the misspelled class regsiterCol to registerCol for coherency.