Error messages
I have removed the commented out error messages for the login attempts of the user from when we were testing the login functionality. I have also implemented the changes that @WS350074 requested for the error message not having the underscores using the str-replace function on the error message being displayed in login.php where the error message is outputted to the user.