Tables on load up and Navbar styling
The main issue I tackled during this commit was getting it so that the session values were working as intended and that a table will display data on load up.
I did this by creating a function which ran a SQL routine in a stmt sequence.
I also added a profile picture to the navbar using the gravatar; for this I had to create a email session value.