further functionality
I have added in the code needed to gather the current date and time of the user when they are logging in and setting it in the database. I have also added in an extra check for the if statement to ensure that the users account state is active otherwise it will not allow access, displaying an error message of their account has been disabled.