Create Class + Styling
I have created a button which opens up a modal. On that modal is a form with an input and a dropdown which gets course name via SQL. This form is able to create new users using an ajax.

I have also done styling for the page making it fit in with the student management.